+++ conda/common/disk.py
@@ -2,11 +2,14 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import errno
+import shutil
 import sys
+from conda import CondaError
 from errno import EACCES, EEXIST, ENOENT, EPERM
 from itertools import chain
 from logging import getLogger
-from os import W_OK, access, chmod, getpid, listdir, lstat, makedirs, rename, unlink, walk
+from os import (W_OK, access, chmod, getpid, link as os_link, listdir, lstat, makedirs, readlink,
+                rename, symlink, unlink, walk, stat)
 from os.path import abspath, basename, dirname, isdir, isfile, islink, join, lexists
 from shutil import rmtree
 from stat import S_IEXEC, S_IMODE, S_ISDIR, S_ISLNK, S_ISREG, S_IWRITE
@@ -14,6 +17,7 @@ from time import sleep
 from uuid import uuid4
 
 from ..compat import lchmod, text_type
+from ..exceptions import CondaOSError
 from ..utils import on_win
 
 __all__ = ["rm_rf", "exp_backoff_fn", "try_write"]
@@ -310,3 +314,115 @@ def yield_lines(path):
             raise StopIteration
         else:
             raise
+
+
+
+
+
+
+
+
+
+
+if on_win:
+    import ctypes
+    from ctypes import wintypes
+
+    CreateHardLink = ctypes.windll.kernel32.CreateHardLinkW
+    CreateHardLink.restype = wintypes.BOOL
+    CreateHardLink.argtypes = [wintypes.LPCWSTR, wintypes.LPCWSTR,
+                               wintypes.LPVOID]
+    try:
+        CreateSymbolicLink = ctypes.windll.kernel32.CreateSymbolicLinkW
+        CreateSymbolicLink.restype = wintypes.BOOL
+        CreateSymbolicLink.argtypes = [wintypes.LPCWSTR, wintypes.LPCWSTR,
+                                       wintypes.DWORD]
+    except AttributeError:
+        CreateSymbolicLink = None
+
+    def win_hard_link(src, dst):
+        "Equivalent to os.link, using the win32 CreateHardLink call."
+        if not CreateHardLink(dst, src, None):
+            raise CondaOSError('win32 hard link failed')
+
+    def win_soft_link(src, dst):
+        "Equivalent to os.symlink, using the win32 CreateSymbolicLink call."
+        if CreateSymbolicLink is None:
+            raise CondaOSError('win32 soft link not supported')
+        if not CreateSymbolicLink(dst, src, isdir(src)):
+            raise CondaOSError('win32 soft link failed')
+
+    def win_conda_bat_redirect(src, dst, shell):
+        """Special function for Windows XP where the `CreateSymbolicLink`
+        function is not available.
+
+        Simply creates a `.bat` file at `dst` which calls `src` together with
+        all command line arguments.
+
+        Works of course only with callable files, e.g. `.bat` or `.exe` files.
+        """
+        from conda.utils import shells
+        try:
+            makedirs(dirname(dst))
+        except OSError as exc:  # Python >2.5
+            if exc.errno == errno.EEXIST and isdir(dirname(dst)):
+                pass
+            else:
+                raise
+
+        # bat file redirect
+        if not isfile(dst + '.bat'):
+            with open(dst + '.bat', 'w') as f:
+                f.write('@echo off\ncall "%s" %%*\n' % src)
+
+        # TODO: probably need one here for powershell at some point
+
+        # This one is for bash/cygwin/msys
+        # set default shell to bash.exe when not provided, as that's most common
+        if not shell:
+            shell = "bash.exe"
+
+        # technically these are "links" - but islink doesn't work on win
+        if not isfile(dst):
+            with open(dst, "w") as f:
+                f.write("#!/usr/bin/env bash \n")
+                if src.endswith("conda"):
+                    f.write('%s "$@"' % shells[shell]['path_to'](src+".exe"))
+                else:
+                    f.write('source %s "$@"' % shells[shell]['path_to'](src))
+            # Make the new file executable
+            # http://stackoverflow.com/a/30463972/1170370
+            mode = stat(dst).st_mode
+            mode |= (mode & 292) >> 2    # copy R bits to X
+            chmod(dst, mode)
+
+
+LINK_HARD = 1
+LINK_SOFT = 2
+LINK_COPY = 3
+link_name_map = {
+    LINK_HARD: 'hard-link',
+    LINK_SOFT: 'soft-link',
+    LINK_COPY: 'copy',
+}
+
+
+def link(src, dst, linktype=LINK_HARD):
+    if linktype == LINK_HARD:
+        if on_win:
+            win_hard_link(src, dst)
+        else:
+            os_link(src, dst)
+    elif linktype == LINK_SOFT:
+        if on_win:
+            win_soft_link(src, dst)
+        else:
+            symlink(src, dst)
+    elif linktype == LINK_COPY:
+        # copy relative symlinks as symlinks
+        if not on_win and islink(src) and not readlink(src).startswith('/'):
+            symlink(readlink(src), dst)
+        else:
+            shutil.copy2(src, dst)
+    else:
+        raise CondaError("Did not expect linktype=%r" % linktype)
