+++ conda/gateways/disk/delete.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+from logging import getLogger
+
+log = getLogger(__name__)
